requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
ftfy==6.1.3
aiohttp==3.9.1
python-multipart==0.0.6
